requirements.txt
sentence-transformers 
numpy
wikipedia
faiss-cpu 
rouge-score
nltk 
tqdm